requirements.txt
gensim
absl-py==0.9.0
astor==0.8.1
blis==0.4.1
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
cymem==2.0.3
de-core-news-sm==2.2.5
en-core-web-sm==2.2.5
enum34==1.1.6
future==0.18.2
gast==0.2.2
google-pasta==0.1.8
grpcio==1.27.1
h5py==2.10.0
idna==2.8
importlib-metadata==1.5.0
joblib==0.14.1
kenlm==0.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2
murmurhash==1.0.2
nltk==3.4.5
normalise==0.1.8
numpy==1.18.1
opt-einsum==3.1.0
pandas==1.0.1
plac==1.1.3
plumbum==1.6.8
preshed==3.0.2
protobuf==3.11.3
pudb==2019.2
Pygments==2.5.2
python-dateutil==2.8.1
PyTreex==0.0.0
pytz==2019.3
PyYAML==5.3
recordclass==0.13.2
regex==2020.1.8
requests==2.22.0
roman==3.2
rpyc==4.1.4
scikit-learn==0.22.1
scipy==1.4.1
sentencepiece==0.1.85
six==1.14.0
spacy==2.2.3
sqlparse==0.3.0
srsly==1.0.1
tensorboard==1.15.0
tensorflow==1.15.2
tensorflow-estimator==1.15.1
termcolor==1.1.0
Theano==1.0.4+unknown
thinc==7.3.1
torch==1.4.0
torchtext==0.5.0
tqdm==4.42.1
unicodecsv==0.14.1
Unidecode==1.1.1
urllib3==1.25.8
urwid==2.1.0
wasabi==0.6.0
Werkzeug==1.0.0
wrapt==1.11.2
zipp==2.1.0
